+++ tests/testthat/test-utils.R
@@ -1,0 +1,7 @@
+context("utils.R")
+
+describe("captures", {
+  it("works with non-ASCII inputs (#120)", {
+    expect_equal(captures("ë", regexpr("(.)", "ë", perl = TRUE))[[1]], "ë")
+  })
+})